+++ Fourmis/egg.cpp
@@ -1,7 +1,7 @@
 #include "egg.h"
 #include "idleegg.h"
 
-Egg::Egg(Environnement& _env, Anthill* _anthill, bool _is_queen, int _max_food_need, int _time_to_transition, int _food_need) :
+Egg::Egg(Environnement& _env, Anthill* _anthill, bool _is_queen, int _max_food_need, int _time_to_transition, float _food_need) :
     Ant(_env, _anthill, _is_queen, _max_food_need, _time_to_transition,_food_need, std::make_unique<IdleEgg>())
 {
