+++ Fourmis/queen.h
@@ -6,7 +6,7 @@ class Queen : public DependentAnt
 {
 public:
     Queen() = delete;
-    Queen(Environnement& _env, Anthill* _anthill, bool _is_queen = false, int _max_food_need = 0, int _time_to_transition = 10, int _food_need = 0, bool _is_being_fed = false);
+    Queen(Environnement& _env, Anthill* _anthill, bool _is_queen = false, int _max_food_need = 0, int _time_to_transition = 10, float _food_need = 0, bool _is_being_fed = false);
     ~Queen();
 
     // Ant interface
